Adding a Link to Club Management
Now a link to "Manage Club" shows up if the user has an organizes
relationship with at least one club.  The clubManagement page still
needs to be developed, needs to show the clubs that the user manages and
needs to allow the user to manage the clubs, so edit information and
create events and edit events.